Fix x86_64 build-install compliance
  - Fix hardcoded library path in CMakeLists.txt
  - Use %cmake macro to set default path
  - Change library path var in *.pc.in template file

+++ src/com-core_packet-router.c
@@ -575,7 +575,7 @@ static void *send_main(void *data)
 		packet_destroy(packet);
 	}
 
-	return (void *)ret;
+	return (void *)(unsigned long)ret;
 }
 
 /*!
@@ -1212,7 +1212,7 @@ static void *server_main(void *data)
 	ctx.pid = (pid_t)-1;
 
 	ret = router_common_main(router, client->handle, &ctx);
-	return (void *)ret;
+	return (void *)(unsigned long)ret;
 }
 
 /*!
@@ -1232,7 +1232,7 @@ static void *client_main(void *data)
 	ctx.pid = (pid_t)-1;
 
 	ret = router_common_main(router, router->handle, &ctx);
-	return (void *)ret;
+	return (void *)(unsigned long)ret;
 }
 
 /*!
